# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Mathia API
  description: |
    API REST pour **Mathia** - Application de révision de mathématiques pour collégiens (6ème à 3ème).
    
    ## 🔐 Authentification
    
    L'API utilise **JWT (JSON Web Tokens)** pour l'authentification.
    
    - Routes **publiques** : `/auth/signup`, `/auth/login`
    - Routes **protégées** : Toutes les autres routes nécessitent un JWT
    
    Pour accéder aux routes protégées, incluez le token dans le header :
    ```
    Authorization: Bearer {votre_token_jwt}
    ```
    
    ## 🚀 Workflow typique
    
    1. S'inscrire avec `/auth/signup` → Récupérer le token
    2. Utiliser le token pour accéder aux ressources protégées
    3. Créer des exercices via `/admin/exercises`
    4. Lister les exercices avec `/exercises`
    5. Soumettre des réponses avec `/attempts`
    
  version: 1.0.0
  contact:
    name: Support Mathia
    email: support@mathia.app
servers:
  - url: http://localhost:3000/api
    description: Serveur de développement local
  - url: https://api.mathia.app/api
    description: Serveur de production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token JWT obtenu via `/auth/signup` ou `/auth/login`.
        
        Format: `Bearer {token}`
  schemas:
...
    Exercise:
      type: object
      properties:
        id:
          type: integer
          example: 1
        courseId:
          type: integer
          example: 1
          description: ID du cours auquel appartient l'exercice
        type:
          type: string
          example: "qcm"
          description: "Type d'exercice (qcm, libre, vrai-faux, calcul)"
        body:
          type: string
          example: "Quelle est la moitié de 10?"
          description: Énoncé de l'exercice
        options:
          type: object
          nullable: true
          example: { "A": "3", "B": "5", "C": "7", "D": "10" }
          description: Options pour les QCM (null pour les autres types)
        answer:
          type: string
          example: "B"
          description: Réponse correcte
        explanation:
          type: string
          nullable: true
          example: "10 divisé par 2 égale 5"
          description: Explication de la solution
        difficulty:
          type: string
          nullable: true
          example: "facile"
          description: "Niveau de difficulté (facile, moyen, difficile)"
        tags:
          type: array
          nullable: true
          items:
            type: string
          example: ["fractions", "division"]
          description: Tags pour catégoriser l'exercice
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        course:
          type: object
          properties:
            id:
              type: integer
            title:
              type: string
            grade:
              type: string
            chapter:
              type: string
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Message d'erreur"
paths:
...
  /exercises:
    get:
      tags:
        - Exercises
      summary: Récupérer tous les exercices avec filtrage
      description: |
        Retourne la liste des exercices avec possibilité de filtrer par :
        - **courseId** : ID du cours
        - **difficulty** : Niveau de difficulté (facile, moyen, difficile)
        
        Les deux filtres peuvent être combinés.
        
        **JWT requis** - Incluez le token dans le header Authorization
      security:
        - bearerAuth: []
      parameters:
        - name: courseId
          in: query
          required: false
          description: Filtrer par ID de cours
          schema:
            type: integer
          example: 1
        - name: difficulty
          in: query
          required: false
          description: Filtrer par niveau de difficulté
          schema:
            type: string
            enum: [facile, moyen, difficile]
          example: "facile"
      responses:
        '200':
          description: Liste des exercices récupérée avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      exercises:
                        type: array
                        items:
                          $ref: '#/components/schemas/Exercise'
              examples:
                all_exercises:
                  summary: Tous les exercices
                  value:
                    success: true
                    data:
                      exercises:
                        - id: 1
                          courseId: 1
                          type: "qcm"
                          body: "Quelle est la moitié de 10?"
                          options: { "A": "3", "B": "5", "C": "7", "D": "10" }
                          answer: "B"
                          explanation: "10 divisé par 2 égale 5"
                          difficulty: "facile"
                          tags: ["fractions", "division"]
                          createdAt: "2025-01-15T10:15:00Z"
                          updatedAt: "2025-01-15T10:15:00Z"
                          course:
                            id: 1
                            title: "Les fractions"
                            grade: "6ème"
                            chapter: "Arithmétique"
                filtered_by_course:
                  summary: Filtrés par cours
                  value:
                    success: true
                    data:
                      exercises:
                        - id: 1
                          courseId: 1
                          type: "qcm"
                          body: "Quelle est la moitié de 10?"
                          difficulty: "facile"
                          course:
                            id: 1
                            title: "Les fractions"
                filtered_by_difficulty:
                  summary: Filtrés par difficulté
                  value:
                    success: true
                    data:
                      exercises:
                        - id: 1
                          difficulty: "facile"
                          body: "Question facile..."
                        - id: 3
                          difficulty: "facile"
                          body: "Autre question facile..."
        '401':
          description: Non authentifié
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /admin/exercises:
    post:
      tags:
        - Admin
      summary: Ajouter un exercice manuellement
      description: |
        Créer un nouvel exercice et l'associer à un cours existant.
        
        **Champs obligatoires :**
        - courseId
        - type
        - body
        - answer
        
        **Champs optionnels :**
        - options (obligatoire pour les QCM)
        - explanation
        - difficulty
        - tags
        
        **JWT requis** - Accessible à tous les utilisateurs authentifiés
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - courseId
                - type
                - body
                - answer
              properties:
                courseId:
                  type: integer
                  example: 1
                  description: ID du cours parent
                type:
                  type: string
                  example: "qcm"
                  description: "Type d'exercice (qcm, libre, vrai-faux, calcul)"
                body:
                  type: string
                  example: "Quelle est la moitié de 10?"
                  description: Énoncé de l'exercice
                options:
                  type: object
                  nullable: true
                  example: { "A": "3", "B": "5", "C": "7", "D": "10" }
                  description: Options pour les QCM (null pour les autres types)
                answer:
                  type: string
                  example: "B"
                  description: Réponse correcte
                explanation:
                  type: string
                  nullable: true
                  example: "10 divisé par 2 égale 5"
                  description: Explication de la solution
                difficulty:
                  type: string
                  nullable: true
                  example: "facile"
                  description: "Niveau de difficulté (facile, moyen, difficile)"
                tags:
                  oneOf:
                    - type: array
                      items:
                        type: string
                    - type: object
                  nullable: true
                  example: ["fractions", "division"]
                  description: Tags pour catégoriser
            examples:
              qcm_exercise:
                summary: Exercice QCM
                value:
                  courseId: 1
                  type: "qcm"
                  body: "Quelle est la moitié de 10?"
                  options:
                    A: "3"
                    B: "5"
                    C: "7"
                    D: "10"
                  answer: "B"
                  explanation: "10 divisé par 2 égale 5"
                  difficulty: "facile"
                  tags: ["fractions", "division"]
              free_exercise:
                summary: Exercice libre
                value:
                  courseId: 1
                  type: "libre"
                  body: "Calculez : 15 + 27"
                  answer: "42"
                  explanation: "15 + 27 = 42"
                  difficulty: "facile"
                  tags: ["addition"]
              minimal_exercise:
                summary: Exercice minimal (champs requis seulement)
                value:
                  courseId: 1
                  type: "calcul"
                  body: "Combien font 5 × 3?"
                  answer: "15"
      responses:
        '201':
          description: Exercice créé avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Exercice créé avec succès"
                  data:
                    type: object
                    properties:
                      exercise:
                        $ref: '#/components/schemas/Exercise'
              examples:
                success:
                  summary: Création réussie
                  value:
                    success: true
                    message: "Exercice créé avec succès"
                    data:
                      exercise:
                        id: 5
                        courseId: 1
                        type: "qcm"
                        body: "Quelle est la moitié de 10?"
                        options: { "A": "3", "B": "5", "C": "7", "D": "10" }
                        answer: "B"
                        explanation: "10 divisé par 2 égale 5"
                        difficulty: "facile"
                        tags: ["fractions", "division"]
                        createdAt: "2025-01-15T14:00:00Z"
                        updatedAt: "2025-01-15T14:00:00Z"
        '400':
          description: Données invalides
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                validation_error:
                  summary: Erreur de validation
                  value:
                    success: false
                    message: "Erreur de validation"
                    errors:
                      - field: "courseId"
                        message: "courseId is required"
        '401':
          description: Non authentifié
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Cours non trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                course_not_found:
                  summary: Cours inexistant
                  value:
                    success: false
                    message: "Cours non trouvé"